# backend/public/openapi.yaml
openapi: 3.1.0
info:
  title: Cultural Site API
  version: 1.0.0
  description: API for managing cultural sites, proposals, reviews, and user data.
servers:
  - url: http://localhost:5000/api/v1
    description: Local Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CulturalSite:
      type: object
      properties:
        _id:
          type: string
          description: Unique identifier for the cultural site.
        name:
          type: string
          description: Name of the cultural site.
          minLength: 2
          maxLength: 100
        description:
          type: string
          description: Description of the cultural site.
          maxLength: 1000
        category:
          type: string
          description: Category of the cultural site.
          enum:
            - artwork
            - gallery
            - museum
            - restaurant
            - theatre
            - arts_centre
            - community_centre
            - library
            - cinema
            - other
        location:
          type: object
          properties:
            type:
              type: string
              enum: ['Point']
              description: GeoJSON type.
            coordinates:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 2
              description: |
                [longitude, latitude] coordinates. Longitude must be between -180 and 180,
                latitude must be between -90 and 90.
          required:
            - type
            - coordinates
        address:
          type: string
          description: Address of the cultural site.
        website:
          type: string
          format: url
          description: Website URL of the cultural site.
        imageUrl:
          type: string
          format: url
          description: Image URL of the cultural site.
        openingHours:
          type: string
          description: Opening hours of the cultural site.
        licenseInfo:
          type: string
          default: 'Data © OpenStreetMap contributors, ODbL.'
          description: Licensing information for the data.
        sourceId:
          type: string
          description: OSM ID (node/way/relation ID) from Overpass API. (node/1234567890)
        reviews:
          type: array
          items:
            type: string
            description: ObjectId of a Review.
        favoritesCount:
          type: number
          format: int32
          minimum: 0
          default: 0
          description: Number of users who favorited this site.
        originalTags:
          type: object
          description: Original GeoJSON 'tags' object from the source element from Overpass API response.
          additionalProperties: true
        proposedBy:
          type: string
          description: ObjectId of the User who proposed this site.
        registeredBy:
          type: string
          description: ObjectId of the User (admin) who registered this site.
        active:
          type: boolean
          default: true
          description: Field for soft deletion (not currently implemented).
      required:
        - name
        - category
        - location
        - address
        - sourceId
        - originalTags
    Review:
      type: object
      properties:
        _id:
          type: string
          description: MongoDB ObjectId for the review.
        culturalSite:
          type: string
          description: ObjectId of the CulturalSite being reviewed.
        user:
          type: string
          description: ObjectId of the User who wrote the review.
        rating:
          type: number
          format: int32
          minimum: 1
          maximum: 5
          description: Rating given by the user (1 to 5).
        comment:
          type: string
          maxLength: 500
          description: Comment provided by the user.
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the review was created.
        active:
          type: boolean
          default: true
          description: Field for soft deletion. (not implemented)
      required:
        - culturalSite
        - user
        - rating
    User:
      type: object
      properties:
        _id:
          type: string
          description: Unique identifier for the user.
        username:
          type: string
          maxLength: 20
          description: User's chosen username.
        email:
          type: string
          format: email
          description: User's email address.
        googleId:
          type: string
          description: Google ID if authenticated via Google.
        profileImage:
          type: string
          format: url
          description: URL to the user's profile image.
        role:
          type: string
          enum: ['user', 'admin']
          default: 'admin'
          description: User's role in the system. temporarily default value is 'admin' to demonstrate the functions.
        active:
          type: boolean
          default: true
          description: Field for soft deletion.
        favoriteSites:
          type: array
          items:
            type: string
            description: ObjectId of a CulturalSite.
          description: List of cultural sites marked as favorites by the user.
        bio:
          type: string
          maxLength: 200
          description: User's short biography.
      required:
        - email
        - googleId
...
paths:
...
  /cultural-sites:
    get:
      summary: Get all cultural sites
      tags:
        - Cultural Sites
      operationId: getAllCulturalSites
      parameters:
        - in: query
          name: category
          schema:
            type: string
          description: Filter by cultural site category.
        - in: query
          name: name
          schema:
            type: string
          description: Filter by cultural site name (partial match).
      responses:
        200:
          description: A list of cultural sites.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CulturalSite'
        500:
          description: Internal server error.
    post:
      summary: Save a cultural site to the database
      tags:
        - Cultural Sites
      operationId: saveCulturalSiteToDb
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CulturalSite'
            examples:
              newCulturalSite:
                value:
                  name: 'Example Museum'
                  description: 'A wonderful museum of modern art.'
                  category: 'Museum'
                  location:
                    type: 'Point'
                    coordinates: [12.345, 54.321]
                  address: '123 Museum St, City'
                  website: 'http://examplemuseum.com'
                  sourceId: 'osm_node_12345'
                  originalTags:
                    amenity: 'museum'
                    name: 'Example Museum'
      responses:
        201:
          description: Cultural site successfully saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CulturalSite'
        400:
          description: Invalid input (e.g., missing fields, invalid location, not in Chemnitz).
        401:
          description: Unauthorized.
        403:
          description: Forbidden (only admins can add sites, or sourceId is in exclusion list).
        409:
          description: Conflict (e.g., sourceId already registered).
        500:
          description: Internal server error during saving.
...
  /users/me:
    get:
      summary: Get current user's profile
      tags:
        - Users
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        200:
          description: Current user's profile details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          description: Unauthorized.
        500:
          description: Internal server error.
    patch:
      summary: Update current user's profile
      tags:
        - Users
      operationId: updateMe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  maxLength: 20
                profileImage:
                  type: string
                bio:
                  type: string
                  maxLength: 200
            examples:
              updateProfile:
                value:
                  username: 'NewUsername'
                  bio: 'My updated bio.'
      responses:
        200:
          description: User profile updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Invalid input.
        401:
          description: Unauthorized.
        500:
          description: Internal server error.
    delete:
      summary: Delete current user's account
      tags:
        - Users
      operationId: deleteMe
      security:
        - bearerAuth: []
      responses:
        200:
          description: User account successfully deleted.
        401:
          description: Unauthorized.
        404:
          description: can't find the user.
        500:
          description: Internal server error.
  /users/{userId}:
    get:
      summary: Get user information by ID
      tags:
        - Users
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: Unique ID of the user.
      responses:
        200:
          description: User profile details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          description: Unauthorized.
        404:
          description: User not found.
        500:
          description: Internal server error.
  /users:
    get:
      summary: Get all user information (Admin only)
      tags:
        - Users
      operationId: getAllUsers
      security:
        - bearerAuth: []
      responses:
        200:
          description: A list of all users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        401:
          description: Unauthorized.
        403:
          description: Forbidden, only admins can view all users.
        500:
          description: Internal server error.
  /users/me/favorites:
    get:
      summary: Get current user's favorite cultural sites
      tags:
        - Users
      operationId: getFavoriteSites
      security:
        - bearerAuth: []
      responses:
        200:
          description: A list of favorite cultural sites.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CulturalSite'
        401:
          description: Unauthorized.
        404:
          description: can't find the user.
        500:
          description: Internal server error.
  /users/me/favorites/{siteId}:
    post:
      summary: Add a cultural site to current user's favorites
      tags:
        - Users
      operationId: addFavoriteSite
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: siteId
          schema:
            type: string
          required: true
          description: Unique ID of the cultural site to add to favorites.
      responses:
        200:
          description: Cultural site added to favorites.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Cultural site already in favorites.
        401:
          description: Unauthorized.
        404:
          description: Cultural site not found.
        500:
          description: Internal server error.
    delete:
      summary: Remove a cultural site from current user's favorites
      tags:
        - Users
      operationId: removeFavoriteSite
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: siteId
          schema:
            type: string
          required: true
          description: Unique ID of the cultural site to remove from favorites.
      responses:
        200:
          description: Cultural site removed from favorites.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Cultural site not in favorites.
        401:
          description: Unauthorized.
        404:
          description: Cultural site not found.
        500:
          description: Internal server error.
  /users/me/reviews:
    get:
      summary: Get reviews submitted by the current user
      tags:
        - Users
      operationId: getMyReviews
      security:
        - bearerAuth: []
      responses:
        200:
          description: A list of reviews submitted by the user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        401:
          description: Unauthorized.
        500:
          description: Internal server error.